test tiny_vec
