+++ tests/test_channel.cpp
@@ -133,29 +133,6 @@ TEST_F(CATEGORY, push_single_threaded) {
   }());
 }
 
-struct destructor_counter {
-  std::atomic<size_t>* count;
-  destructor_counter(std::atomic<size_t>* C) noexcept : count{C} {}
-  destructor_counter(destructor_counter const& Other) = delete;
-  destructor_counter& operator=(destructor_counter const& Other) = delete;
-
-  destructor_counter(destructor_counter&& Other) noexcept {
-    count = Other.count;
-    Other.count = nullptr;
-  }
-  destructor_counter& operator=(destructor_counter&& Other) noexcept {
-    count = Other.count;
-    Other.count = nullptr;
-    return *this;
-  }
-
-  ~destructor_counter() {
-    if (count != nullptr) {
-      ++(*count);
-    }
-  }
-};
-
 TEST_F(CATEGORY, destroy_chan_with_data) {
   test_async_main(ex(), []() -> tmc::task<void> {
     std::atomic<size_t> count;
